Safari-609.1.11.1/LayoutTests/fast/animation/request-animation-frame-too-rapid-expected.txt

Tests that requestAnimationFrame callbacks aren't called too rapidly (above 1000FPS)

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
PASS 100 is >= numCallbacks
